ghdd

+++ EmployeeAttachment.Application/Features/Employee/Commands/UpdateEmployee/UpdateEmployeeCommand.cs
@@ -14,7 +14,7 @@ namespace EmployeeAttachment.Application.Features.Employee.Commands.UpdateEmploy
         public string? LastName { get; set; }
         public string? Email { get; set; }
         public string? Image { get; set; }
-        public decimal? Salary { get; set; }
+        public int? Salary { get; set; }
         public string? Phone { get; set; }
         public string? FullName { get; set; }
         public DateTime HireDate { get; set; }
